불필요한 코드 제거, 기존 Mock ControllerTest 이름 변경, 통합테스트 작성
ProductController의 통합 테스트를 만들었습니다.
생성 메서드를 실패, 빠른 성공 사이클을 수행했습니다.
